requirements.txt
# Required core packages for the advanced query engine
crewai==0.28.8
langchain==0.3.24
langchain-core==0.3.56
langchain-openai==0.3.14
serpapi==0.1.4
tavily-python==0.4.0
requests==2.32.3
python-dotenv==1.1.0
openai==1.76.0

# Other dependencies
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
autobahn==19.11.2
Automat==25.4.16
backoff==2.2.1
banks==2.1.2
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==1.0.7
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
constantly==23.10.4
cryptography==44.0.2
dash==3.0.4
dash-bootstrap-components==2.0.2
dataclasses-json==0.6.7
dateparser==1.2.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
durationpy==0.9
ecdsa==0.19.1
et_xmlfile==2.0.0
fastapi==0.115.9
filelock==3.18.0
filetype==1.2.0
Flask==3.0.3
flatbuffers==25.2.10
frozendict==2.4.6
frozenlist==1.6.0
fsspec==2025.3.2
google-auth==2.39.0
googleapis-common-protos==1.70.0
greenlet==3.2.1
griffe==1.7.3
grpcio==1.71.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.30.2
humanfriendly==10.0
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
incremental==24.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiteconnect==5.0.1
kubernetes==32.0.1
langchain-text-splitters==0.3.8
langsmith==0.3.37
llama-cloud==0.1.19
llama-cloud-services==0.6.16
llama-index==0.12.33
llama-index-agent-openai==0.4.6
llama-index-cli==0.4.1
llama-index-core==0.12.33.post1
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.11
llama-index-llms-openai==0.3.38
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.7
llama-index-readers-llama-parse==0.4.0
llama-parse==0.6.16
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.4.3
multitasking==0.0.11
mypy_extensions==1.1.0
narwhals==1.36.0
nest-asyncio==1.6.0
networkx==3.4.2
newsapi==0.1.1
newsapi-python==0.2.7
nltk==3.9.1
numpy==2.2.5
oauthlib==3.2.2
onnxruntime==1.21.1
openpyxl==3.1.5
opentelemetry-api==1.32.1
opentelemetry-exporter-otlp-proto-common==1.32.1
opentelemetry-exporter-otlp-proto-grpc==1.32.1
opentelemetry-instrumentation==0.53b1
opentelemetry-instrumentation-asgi==0.53b1
opentelemetry-instrumentation-fastapi==0.53b1
opentelemetry-proto==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
opentelemetry-util-http==0.53b1
orjson==3.10.16
overrides==7.7.0
packaging==24.2
pandas==2.2.3
passlib==1.7.4
peewee==3.17.9
pillow==11.2.1
platformdirs==4.3.7
plotly==6.0.1
posthog==4.0.0
propcache==0.3.1
protobuf==5.29.4
pyasn1==0.4.8
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
pyOpenSSL==25.0.0
pypdf==5.4.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-binance==1.0.28
python-dateutil==2.9.0.post0
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.2
pywin32==310
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retrying==1.3.4
rich==14.0.0
rpds-py==0.24.0
rsa==4.9.1
service-identity==24.2.0
setuptools==79.0.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
starlette==0.45.3
striprtf==0.0.26
sympy==1.13.3
ta==0.11.0
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.1
tqdm==4.67.1
Twisted==24.11.0
txaio==23.1.1
typer==0.15.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
uvicorn==0.34.2
watchdog==6.0.0
watchfiles==1.0.5
websocket-client==1.8.0
websockets==15.0.0
Werkzeug==3.0.6
wrapt==1.17.2
yarl==1.20.0
yfinance==0.2.56
zipp==3.21.0
zope.interface==7.2
zstandard==0.23.0
